added patch and delete request and edited models to include image property
