+++ models/anime.js
@@ -3,7 +3,8 @@ const mongoose = require('../db/connection')
 const animeSchema = new mongoose.Schema({
     name: String,
     yearReleased: String,
-    genre: String
+    genre: String,
+    image: String
 
 })
