--- models/user.js
@@ -1,7 +1,0 @@
-const mongoose = require("../db/connection")
-
-const userSchema = new mongoose.Schema({
-    name: String
-})
-
-module.exports = mongoose.model('User', userSchema)
