+++ server.js
@@ -2,15 +2,16 @@ const express = require('express')
 const cors = require('cors')
 require('dotenv').config()
 const app = express()
-const animeController = require('./controllers/animeController')
-const userController = require('./controllers/userController')
-
 const PORT = process.env.PORT
+
+const animeController = require('./controllers/animeController')
+const animeMovieController = require('./controllers/animeMovieController')
+
 app.use(cors())
 app.use(express.json())
 
 app.use('/animes', animeController)
-app.use('/users', userController)
+app.use('/animeMovies', animeMovieController)
